OpenMP build option and windows build fixes
Note : openmp is really bad on windows but it is now supported.
it requires using only signed int in for loops and unary operators
on ints for atomics.

+++ src/Tests/CoreTests/Manager.cpp
@@ -37,7 +37,7 @@ namespace RaTests {
             }
         }
 
-        printf("Result : %lu / %lu tests passed\n", m_tests.size() - numFailed, m_tests.size());
+        printf("Result : %lu / %lu tests passed\n", ulong(m_tests.size() - numFailed), ulong(m_tests.size()));
         return numFailed;
     }
